refactor(layout): extract table detection
Move layout table detection and numeric cell normalization into a dedicated module so the layout entrypoint only coordinates table extraction.
